Introduce pieRecipe to control topping decision
Improve the AddSprinklesToPie feature flag by using the dependency injection pattern to invert the decision of which topping to use.
Instead of having an "if"-statement in the middle of the BakePie method, the function createFeatureAwareRecipeAsync is used to create a recipe that takes feature flag(s) from FeatureHub into consideration and controls which topping the pie will have.